mathlearning/requirements.txt
sympy==1.4
requests
flask-cors
antlr4-python3-runtime
Django
djangorestframework